requirements.txt
PySide6
mutagen
pyacoustid
musicbrainzngs
